Fix format string issues and add safety checks for ThreadPoolManager
- Replace %zu format specifiers with %u and proper casting to fix [ERRFMT] errors
- Add safety checks for type string corruption in ThreadPoolManager
- Make copy of round_robin_types_ vector to avoid modification during iteration
- Add hex debugging output to trace type string corruption
- Add safety checks for empty/invalid type strings in logging

+++ src/orchestration/scheduler_service.cpp
@@ -33,7 +33,7 @@ namespace MediaDedup::Orchestration
         running_.store(false, std::memory_order_relaxed);
         std::lock_guard<std::mutex> lock(jobsMutex_);
 
-        logger.debug("Stopping %zu registered jobs", jobs_.size());
+        logger.debug("Stopping %u registered jobs", static_cast<unsigned int>(jobs_.size()));
         for (auto &kv : jobs_)
         {
             auto &job = *kv.second;
